🚀 package (uniswap_hft): move all submodules into a `uniswap_hft` named package (#15)
* 🚀 package (uniswap_hft): move all submodules into a `uniswap_hft`name package

* 🖤 Format code with black

* 🚚 chore(app.py): update import path for trading_engine
🚚 chore(test_api.py): update import path for TradingEngine
🚚 chore(test_uniswap_math.py): update import path for TokenManagement
The import paths for the trading_engine, TradingEngine, and TokenManagement modules have been updated to reflect the new package structure. This change was made to improve the organization of the codebase.

* 🖤 Format code with black

* 🐛 fix(test.yml): change package name to uniswap-hft
The package name was incorrect and has been updated to the correct package name, uniswap-hft.

* 🚀 chore(test.yml): remove uniswap_hft/. from pip install command
📦 build(setup.py): add package metadata and dependencies
🧪 test(test_api.py): import TradingEngineAPI from correct module
🔧 refactor(engine.py): fix import statement for web_manager
The pip install command in the test.yml file was modified to remove the uniswap_hft/. path, which was causing issues with the pytest-action. The setup.py file was added to the project to provide metadata about the package and its dependencies. The test_api.py file was modified to import TradingEngineAPI from the correct module. The engine.py file was modified to fix the import statement for web_manager.

* 🖤 Format code with black

* 🔧 chore(test.yml): update python-version to 3.10
The python-version in the matrix strategy has been updated to 3.10 to ensure that the tests are run on the latest version of Python. This change is a routine update to keep the project up-to-date with the latest software.

* 🚀 chore(README.md): update Python version requirement to 3.10
The Python version requirement has been updated to 3.10 to ensure compatibility with the latest version of Python.

* 🚚 chore(api.py, web_manager.py): move engine and uniswap_math imports to correct paths
The imports for `engine` and `TokenManagement` were moved to the correct paths to match the package structure. This ensures that the imports are correct and the code can run without errors.

* 🖤 Format code with black

* 🐛 fix(engine.py): fix index out of range error when position_history is empty
The fix checks if the position_history list is not empty before accessing the last element. This prevents an index out of range error that occurs when the list is empty.

---------
