+++ .github/workflows/test.yml
@@ -8,7 +8,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        python-version: ["3.9"]
+        python-version: ["3.10"]
 
     steps:
       - uses: szenius/set-timezone@v1.2
@@ -20,7 +20,7 @@ jobs:
         with:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
-        run: pip install pytest pytest-md pytest-emoji package/.
+        run: pip install pytest pytest-md pytest-emoji .
       - uses: pavelzw/pytest-action@v2
         with:
           emoji: true
