+++ test/test_uniswap_math.py
@@ -1,6 +1,6 @@
 import math
 
-from uniswap_math import TokenManagement
+from uniswap_hft.uniswap_math import TokenManagement
 
 # Initialize the TokenManager with token0 decimals 6 and token1 decimals 18
 token_manager = TokenManagement.TokenManager(6, 18)
