+++ uniswap_hft/trading_engine/api.py
@@ -6,7 +6,8 @@ from flask import Flask, jsonify, request
 from flask_jwt_extended import (JWTManager, create_access_token,
                                 create_refresh_token, get_jwt_identity,
                                 jwt_required)
-from trading_engine import engine
+
+from uniswap_hft.trading_engine import engine
 
 
 class TradingEngineAPI:
