+++ uniswap_hft/uniswap_v3/assets/erc20.abi
