+++ uniswap_hft/uniswap_v3/assets/erc20b32.abi
